% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableRows.R
\name{getTableRows}
\alias{getTableRows}
\title{Retrieves all rows from Coda table}
\usage{
getTableRows(doc_id, table_id, bearer_token, prod = TRUE)
}
\arguments{
\item{doc_id}{The doc id for the Coda doc (10 characters long)}

\item{table_id}{The table or view id to query}

\item{prod}{Which Coda environment to send API requests to}
}
\description{
Retrieves all rows from Coda table
}
