% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableRowsHelper.R
\name{getTableRowsHelper}
\alias{getTableRowsHelper}
\title{Retrieves a paginated batch of rows and returns tibble}
\usage{
getTableRowsHelper(coda_url, bearer_token)
}
\arguments{
\item{coda_url}{A Coda API request URL for rows}

\item{bearer_token}{A valid Coda API token}
}
\description{
Retrieves a paginated batch of rows and returns tibble
}
